pallet_session_validator_management_rpc

Trait SessionValidatorManagementRpcApiClient

pub trait SessionValidatorManagementRpcApiClient: ClientT {
    // Provided methods
    fn get_epoch_committee<'life0, 'async_trait>(
        &'life0 self,
        epoch_number: u64,
    ) -> Pin<Box<dyn Future<Output = Result<GetCommitteeResponse, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_registrations<'life0, 'async_trait>(
        &'life0 self,
        mc_epoch_number: McEpochNumber,
        stake_pool_public_key: StakePoolPublicKey,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<CandidateRegistrationEntry>, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn get_ariadne_parameters<'life0, 'async_trait>(
        &'life0 self,
        epoch_number: McEpochNumber,
    ) -> Pin<Box<dyn Future<Output = Result<AriadneParameters, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the SessionValidatorManagementRpcApi RPC API.

Provided Methods

fn get_epoch_committee<'life0, 'async_trait>( &'life0 self, epoch_number: u64, ) -> Pin<Box<dyn Future<Output = Result<GetCommitteeResponse, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the committee for given sidechain epoch. The order of the list represents the order of slot allocation.

fn get_registrations<'life0, 'async_trait>( &'life0 self, mc_epoch_number: McEpochNumber, stake_pool_public_key: StakePoolPublicKey, ) -> Pin<Box<dyn Future<Output = Result<Vec<CandidateRegistrationEntry>, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

returns: Last active and valid registration followed by all newer invalid registrations for mc_epoch_number and mc_public_key. Regardless of mc_epoch_number value, it always uses validation api from the latest sidechain block.

fn get_ariadne_parameters<'life0, 'async_trait>( &'life0 self, epoch_number: McEpochNumber, ) -> Pin<Box<dyn Future<Output = Result<AriadneParameters, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Regardless of epoch_number value, all the candidates data validation is done based on the validation api from the latest sidechain block.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.